Introduce dir checks

+++ boot3.sh
@@ -1,10 +1,15 @@
 set -o verbose
 
-mkdir ~/AUR
-cd ~/AUR
-rm -rf preloader-signed
+cd ~
+
+if [ ! -d AUR ]; then mkdir AUR; fi
+cd AUR
+
+if [ -d preloader-signed ]; then rm -rf preloader-signed; fi
 git clone https://aur.archlinux.org/preloader-signed.git
 cd preloader-signed
+
 makepkg -si --noconfirm
+
 cd ~
